+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.7.0] - 2019-08-06
 ### Added
 - `Node::prev_sibling_element` and `Node::next_sibling_element`.
 
@@ -80,7 +82,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `ExpandedName::has_namespace`. `ExpandedName::namespace` will return `Option` now.
 - `Error::NestedEntityReference`.
 
-[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.6.1...HEAD
+[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/RazrFalcon/roxmltree/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.4.1...v0.5.0